Create vector_field_grapher.js
Graphs a vector field in all of 3d space from a function f(x,y,z) and graphs output set {x1,y1,z1} as a vector from the input points. Vectors are artificially shortened and color-coded from blue→green→red (in order of magnitude) using a blending algorithm. Regions of f that are undefined are not graphed by default.